Move KapRulesTable into separate file
